REF: Use `openff-models`
