+++ devtools/conda-envs/beta_env.yaml
@@ -12,6 +12,7 @@ dependencies:
   - openmm >=7.6
   # OpenFF stack
   - openff-toolkit >=0.11.3
+  - openff-models
   # Optional features
   - jax
   - jaxlib
