+++ devtools/conda-envs/dev_env.yaml
@@ -11,6 +11,7 @@ dependencies:
   - openmm
   # OpenFF stack
   - openff-toolkit >=0.11.3
+  - openff-models
   # Optional features
   - jax
   - jaxlib
