+++ devtools/conda-envs/docs_env.yaml
@@ -10,6 +10,7 @@ dependencies:
   - pydantic
   - pint
   - openff-toolkit-base >=0.11.3
+  - openff-models
   - openmm >=7.6
   - intermol
   - jax
