+++ devtools/conda-envs/minimal_env.yaml
@@ -9,6 +9,7 @@ dependencies:
   - pydantic <2.0.0a0
   # OpenFF stack
   - openff-toolkit >=0.11.3
+  - openff-models
   # Testing
   - pytest
   - pytest-cov
