+++ openff/interchange/components/mdconfig.py
@@ -1,6 +1,8 @@
 """Runtime settings for MD simulations."""
 from typing import TYPE_CHECKING, Literal
 
+from openff.models.models import DefaultModel
+from openff.models.types import FloatQuantity
 from openff.units import unit
 from pydantic import Field
 
@@ -9,8 +11,6 @@ from openff.interchange.exceptions import (
     UnsupportedCutoffMethodError,
     UnsupportedExportError,
 )
-from openff.interchange.models import DefaultModel
-from openff.interchange.types import FloatQuantity
 
 if TYPE_CHECKING:
     from openff.interchange import Interchange
