+++ openff/interchange/models.py
@@ -1,25 +1,9 @@
 """Custom Pydantic models."""
 from typing import Optional, Tuple
 
-from openff.units import unit
-from pydantic import BaseModel, Field
+from openff.models.models import DefaultModel
+from pydantic import Field
 from typing_extensions import Literal
-
-from openff.interchange.types import custom_quantity_encoder, json_loader
-
-
-class DefaultModel(BaseModel):
-    """A custom Pydantic model used by other components."""
-
-    class Config:
-        """Custom Pydantic configuration."""
-
-        json_encoders = {
-            unit.Quantity: custom_quantity_encoder,
-        }
-        json_loads = json_loader
-        validate_assignment = True
-        arbitrary_types_allowed = True
 
 
 class TopologyKey(DefaultModel):
